style: RacingCars -> RacingCarList 클래스명 변경
